Error and Input structure refactored
