+++ src/features/lecture-detail/views/homework-item/homework-content/send-homework/send-homework.tsx
@@ -1,19 +1,18 @@
 import { FC, useContext, ChangeEvent } from "react";
-import { FormControl, FormHelperText } from "@mui/material";
 import { SubmitHandler, useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup/dist/yup";
 import * as yup from "yup";
 import { useTranslation } from "react-i18next";
-import RHF from "shared/components/input-RHF";
 import { client } from "api";
 import { LectureIdContext } from "features/lecture-detail/context/lecture-id-context";
+import InputText from "shared/components/form/input-text";
+import { MAX_HOMEWORK_LENGTH } from "shared/constants";
 import { ISendHomeWorkContent, ISendHomeWork } from "./send-homework.types";
 import {
   StyledBox,
   StyledLoadingButton,
   StyledStack,
 } from "./send-homework.styled";
-import { MAX_HOMEWORK_LENGTH } from "../../../../../../shared/constants";
 
 const SendHomework: FC<ISendHomeWork> = (props) => {
   const { sendHomeWorkToCheck, loading } = props;
@@ -59,23 +58,19 @@ const SendHomework: FC<ISendHomeWork> = (props) => {
     <form>
       <StyledStack>
         <StyledBox>
-          <FormControl fullWidth>
-            <RHF.InputTextField
-              placeholder="Текст ответа"
-              multiline
-              maxRows={10}
-              minRows={5}
-              name="content"
-              control={control}
-              inputProps={{
-                maxLength: MAX_HOMEWORK_LENGTH,
-                onChange: handleChange,
-              }}
-            />
-            {errors?.content && (
-              <FormHelperText error>{errors?.content.message}</FormHelperText>
-            )}
-          </FormControl>
+          <InputText
+            placeholder="Текст ответа"
+            multiline
+            maxRows={10}
+            minRows={5}
+            name="content"
+            control={control}
+            inputProps={{
+              maxLength: MAX_HOMEWORK_LENGTH,
+              onChange: handleChange,
+            }}
+            errors={errors}
+          />
           <StyledLoadingButton
             variant="contained"
             loading={loading}
