+++ src/shared/components/form/input-date/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./input-date";
