+++ src/shared/components/form/input-date/input-date.test.tsx
@@ -1,7 +1,7 @@
 import { MemoryRouter } from "react-router-dom";
 import { FormProvider, useForm } from "react-hook-form";
 import { render } from "test/utils-test";
-import InputDatePicker from "./input-date-picker";
+import InputDate from "./input-date";
 
 describe("InputDatePicker", () => {
   it("renders correctly", () => {
@@ -10,7 +10,7 @@ describe("InputDatePicker", () => {
       return (
         <MemoryRouter>
           <FormProvider {...methods}>
-            <InputDatePicker
+            <InputDate
               control={methods.control}
               name="testDate"
               label="testDate"
