+++ src/shared/components/form/input-date/input-date.types.ts
@@ -1,0 +1,8 @@
+import { Control, FieldValues } from "react-hook-form";
+
+export interface IFormInputDate<T extends FieldValues> {
+  control: Control<T, unknown>;
+  label?: string;
+  name: T[keyof T];
+  onChange?: (value: string) => void;
+}
