+++ src/shared/components/form/input-phone/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./input-phone";
