+++ src/shared/components/form/input-phone/input-phone.styled.ts
@@ -1,0 +1,5 @@
+import { styled } from "@mui/system";
+
+export const StyledImgBox = styled("li")({
+  "& > img": { marginRight: "20px", flexShrink: 0 },
+});
