+++ src/shared/components/form/input-phone/input-phone.test.tsx
@@ -1,22 +1,22 @@
 import { MemoryRouter } from "react-router-dom";
-import { FormProvider, useForm } from "react-hook-form";
+import { FieldErrors, FormProvider, useForm } from "react-hook-form";
 import { render } from "test/utils-test";
-import InputTextField from "./input-text-field";
+import InputPhone from "./input-phone";
 
-describe("InputTextField", () => {
+describe("InputPhone", () => {
   it("renders correctly", () => {
     function TestComponent() {
       const methods = useForm();
+      const errors: FieldErrors = {};
       return (
         <MemoryRouter>
           <FormProvider {...methods}>
-            <InputTextField
+            <InputPhone
               control={methods.control}
               name="testDate"
-              placeholder="testDate"
               label="testDate"
-              maxRows="4"
-              minRows="2"
+              placeholder="testDate"
+              errors={errors}
             />
           </FormProvider>
         </MemoryRouter>
