+++ src/shared/components/form/input-radio/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./input-radio";
