+++ src/shared/components/form/input-select/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./input-select";
