+++ src/shared/components/form/input-select/input-select.styled.ts
@@ -1,0 +1,6 @@
+import { styled } from "@mui/system";
+import { InputLabel } from "@mui/material";
+
+export const StyledInputLabel = styled(InputLabel)(({ theme }) => ({
+  backgroundColor: theme.palette.white.main,
+}));
