+++ src/shared/components/form/input-text/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./input-text";
