--- src/shared/components/input-RHF/index.ts
@@ -1,13 +1,0 @@
-import InputTextField from "./input-text-field/input-text-field";
-import InputSelect from "./input-select/input-select";
-import InputPhone from "./input-phone/input-phone";
-import InputDatePicker from "./input-date-picker/input-date-picker";
-import InputRadio from "./input-radio/input-radio";
-
-export default {
-  InputSelect,
-  InputTextField,
-  InputDatePicker,
-  InputPhone,
-  InputRadio,
-};
