+++ src/shared/components/layout/layout.types.ts
@@ -1,4 +1,4 @@
-import { UserRole } from "../../../api/graphql/generated/graphql";
+import { UserRole } from "api/graphql/generated/graphql";
 
 export interface ILayout {
   userRoles: Array<UserRole | null>;
